requirements.txt
flask
firebase_admin
python-dotenv
pyrebase4
requests-toolbelt==0.10.1
Flask-WTF
bootstrap-flask
Gunicorn